Updates cadence to latest versions
