--- cadence/scripts/child_account/get_all_account_balances_from_storage.cdc
@@ -1,121 +1,0 @@
-import FungibleToken from "../../contracts/utility/FungibleToken.cdc"
-import MetadataViews from "../../contracts/utility/MetadataViews.cdc"
-import FungibleTokenMetadataViews from "../../contracts/utility/FungibleTokenMetadataViews.cdc"
-import ChildAccount from "../../contracts/ChildAccount.cdc"
-
-/// Custom struct to easily communicate vault data to a client
-pub struct VaultInfo {
-    pub let name: String?
-    pub let symbol: String?
-    pub var balance: UFix64
-    pub let description: String?
-    pub let externalURL: String?
-    pub let logos: MetadataViews.Medias?
-    pub let storagePathIdentifier: String
-    pub let receiverPathIdentifier: String?
-    pub let providerPathIdentifier: String?
-
-    init(
-        name: String?,
-        symbol: String?,
-        balance: UFix64,
-        description: String?,
-        externalURL: String?,
-        logos: MetadataViews.Medias?,
-        storagePathIdentifier: String,
-        receiverPathIdentifier: String?,
-        providerPathIdentifier: String?
-    ) {
-        self.name = name
-        self.symbol = symbol
-        self.balance = balance
-        self.description = description
-        self.externalURL = externalURL
-        self.logos = logos
-        self.storagePathIdentifier = storagePathIdentifier
-        self.receiverPathIdentifier = receiverPathIdentifier
-        self.providerPathIdentifier = providerPathIdentifier
-    }
-
-    pub fun addBalance(_ addition: UFix64) {
-        self.balance = self.balance + addition
-    }
-}
-
-/// Returns a dictionary of VaultInfo indexed on the Type of Vault
-pub fun getAllVaultInfoInAddressStorage(_ address: Address): {Type: VaultInfo} {
-    // Get the account
-    let account: AuthAccount = getAuthAccount(address)
-    // Init for return value
-    let balances: {Type: VaultInfo} = {}
-    // Assign the type we'll need
-    let vaultType: Type = Type<@{FungibleToken.Balance, MetadataViews.Resolver}>()
-    let ftViewType: Type= Type<FungibleTokenMetadataViews.FTView>()
-    // Iterate over all stored items & get the path if the type is what we're looking for
-    account.forEachStored(fun (path: StoragePath, type: Type): Bool {
-        if type.isSubtype(of: vaultType) {
-            // Get a reference to the vault & its balance
-            if let vaultRef = account.borrow<&{FungibleToken.Balance, MetadataViews.Resolver}>(from: path) {
-                let balance = vaultRef.balance
-                // Attempt to resolve metadata on the vault
-                if let ftView = vaultRef.resolveView(ftViewType) as! FungibleTokenMetadataViews.FTView? {
-                    // Insert a new info struct if it's the first time we've seen the vault type
-                    if !balances.containsKey(type) {
-                        let vaultInfo = VaultInfo(
-                            name: ftView.ftDisplay?.name ?? vaultRef.getType().identifier,
-                            symbol: ftView.ftDisplay?.symbol,
-                            balance: balance,
-                            description: ftView.ftDisplay?.description,
-                            externalURL: ftView.ftDisplay?.externalURL?.url,
-                            logos: ftView.ftDisplay?.logos,
-                            storagePathIdentifier: path.toString(),
-                            receiverPathIdentifier: ftView.ftVaultData?.receiverPath?.toString(),
-                            providerPathIdentifier: ftView.ftVaultData?.providerPath?.toString()
-                        )
-                        balances.insert(key: type, vaultInfo)
-                    } else {
-                        // Otherwise just update the balance of the vault (unlikely we'll see the same type twice in
-                        // the same account, but we want to cover the case)
-                        balances[type]!.addBalance(balance)
-                    }
-                }
-            }
-        }
-        return true
-    })
-    return balances
-}
-
-/// Takes two dictionaries containing VaultInfo structs indexed on the type of vault they represent &
-/// returns a single dictionary containg the summed balance of each respective vault type
-pub fun merge(_ d1: {Type: VaultInfo}, _ d2: {Type: VaultInfo}): {Type: VaultInfo} {
-    for type in d1.keys {
-        if d2.containsKey(type) {
-            d1[type]!.addBalance(d2[type]!.balance)
-        }
-    }
-
-    return d1
-}
-
-pub fun main(address: Address): {Type: VaultInfo} {
-    // Get the balance info for the given address
-    var balances: {Type: VaultInfo} = getAllVaultInfoInAddressStorage(address)
-    
-    /* Iterate over any child accounts */ 
-    //
-    // Get reference to ChildAccountManager if it exists
-    if let managerRef = getAccount(address).getCapability<
-            &{ChildAccount.ChildAccountManagerViewer}
-        >(
-            ChildAccount.ChildAccountManagerPublicPath
-        ).borrow() {
-        // Iterate over each child account in ChildAccountManagerRef
-        for childAddress in managerRef.getChildAccountAddresses() {
-            // Ensure all vault type balances are pooled across all addresses
-            balances = merge(balances, getAllVaultInfoInAddressStorage(childAddress))
-        }
-    }
-    return balances 
-}
-
