--- packages/shared/cadence/scripts/child-account/get-child-address-from-public-key-on-creator.ts
@@ -1,29 +1,0 @@
-const GET_CHILD_ADDRESS_FROM_PUBLIC_KEY_ON_CREATOR = `
-import ChildAccount from 0xChildAccount
-
-/// Returns the child address associated with a public key if account
-/// was created by the ChildAccountCreator at the specified Address and
-/// the provided public key is still active on the account.
-///
-pub fun main(creatorAddress: Address, pubKey: String): Address? {
-  // Get a reference to the ChildAccountCreatorPublic Capability from creatorAddress
-  if let creatorRef = getAccount(creatorAddress)
-    .getCapability<
-      &{ChildAccount.ChildAccountCreatorPublic}
-    >(
-      ChildAccount.ChildAccountCreatorPublicPath
-    ).borrow() {
-		// Get the address created by the given public key if it exists
-    if let address = creatorRef.getAddressFromPublicKey(publicKey: pubKey) {
-			// Also check that the given key has not been revoked
-      if ChildAccount.isKeyActiveOnAccount(publicKey: pubKey, address: address) {
-        return address
-      }
-    }
-    return nil
-  }
-  return nil
-}
-`
-
-export default GET_CHILD_ADDRESS_FROM_PUBLIC_KEY_ON_CREATOR
